Moved BMT.java ahead of gui tests
